requirements.txt
streamlit==1.31.0
google-generativeai
requests
beautifulsoup4
PyPDF2
streamlit-lottie